Turn unconnected socket into a UnixConn
Go by default does not allow you to create an unconnected socket.

You either need to Dial() and get a UnixConn or you need to Listen()
and get a UnixConn

The problem is we don't want to deal with connections as we're simply
sending datagrams. And we don't want to deal with the scenario when
the journal socket is unavailable.

Before I did this by creating an abstract unix domain socket and
writing from there. However MacOS doesn't support abstract sockets.

So instead we call the socket syscall manually and then coerce it
back into a UnixConn.